feature:Add the function of self-selecting rules

+++ app/application/src/main/java/com/alipay/application/service/rule/RuleServiceImpl.java
@@ -29,11 +29,9 @@ import com.alipay.application.service.common.utils.DbCacheUtil;
 import com.alipay.application.service.rule.domain.repo.RuleGroupRepository;
 import com.alipay.application.service.rule.enums.RuleType;
 import com.alipay.application.service.rule.exposed.GroupJoinService;
+import com.alipay.application.service.system.domain.repo.TenantRepository;
 import com.alipay.application.share.request.base.IdRequest;
-import com.alipay.application.share.request.rule.ChangeStatusRequest;
-import com.alipay.application.share.request.rule.ListRuleRequest;
-import com.alipay.application.share.request.rule.RegoRequest;
-import com.alipay.application.share.request.rule.SaveRuleRequest;
+import com.alipay.application.share.request.rule.*;
 import com.alipay.application.share.vo.ApiResponse;
 import com.alipay.application.share.vo.ListVO;
 import com.alipay.application.share.vo.rule.RuleTypeVO;
@@ -64,6 +62,10 @@ public class RuleServiceImpl implements RuleService {
     @Resource
     private RuleMapper ruleMapper;
     @Resource
+    private TenantRuleMapper tenantRuleMapper;
+    @Resource
+    private TenantRepository tenantRepository;
+    @Resource
     private GroupJoinService groupJoinService;
     @Resource
     private RuleRegoMapper ruleRegoMapper;
@@ -83,6 +85,7 @@ public class RuleServiceImpl implements RuleService {
     private DbCacheUtil dbCacheUtil;
 
     public static final String cacheKey = "rule::query_rule_list";
+    public static final String tenantSelectRuleCacheKey = "rule::query_tenant_select_rule_list";
 
     @Transactional(rollbackFor = Exception.class)
     @Override
@@ -205,9 +208,66 @@ public class RuleServiceImpl implements RuleService {
         return new ApiResponse<>(listVO);
     }
 
+    @Override
+    public ListVO<RuleVO> queryTenantSelectRuleList(ListRuleRequest request) {
+        boolean needCache = false;
+        String key = CacheUtil.buildKey(tenantSelectRuleCacheKey, UserInfoContext.getCurrentUser().getUserTenantId(), request.getPage(), request.getSize(),
+                request.getSortParam(), request.getSortType());
+        if (ListUtils.isEmpty(request.getPlatformList())
+                && ListUtils.isEmpty(request.getRuleTypeIdList())
+                && ListUtils.isEmpty(request.getResourceTypeList())
+                && StringUtils.isEmpty(request.getStatus())
+                && ListUtils.isEmpty(request.getRuleCodeList())
+                && ListUtils.isEmpty(request.getRiskLevelList())
+                && ListUtils.isEmpty(request.getRuleGroupIdList())) {
+            needCache = true;
+            DbCachePO dbCachePO = dbCacheUtil.get(key);
+            if (dbCachePO != null) {
+                return JSON.parseObject(dbCachePO.getValue(), new TypeReference<>() {
+                });
+            }
+        }
+
+        RuleDTO ruleDTO = RuleDTO.builder().build();
+        BeanUtils.copyProperties(request, ruleDTO);
+        ruleDTO.setResourceTypeList(ListUtils.setList(request.getResourceTypeList()));
+        ruleDTO.setRuleTypeIdList(ListUtils.setList(request.getRuleTypeIdList()));
+        ruleDTO.setTenantId(UserInfoContext.getCurrentUser().getUserTenantId());
+
+        ListVO<RuleVO> listVO = new ListVO<>();
+        int count = tenantRuleMapper.findCount(ruleDTO);
+        if (count == 0) {
+            return listVO;
+        }
+
+        ruleDTO.setOffset();
+        List<RulePO> list = tenantRuleMapper.findSortList(ruleDTO);
+
+        List<RuleVO> collect = list.stream().map(RuleVO::buildEasy).toList();
+        listVO.setTotal(count);
+        listVO.setData(collect);
+
+        if (needCache) {
+            dbCacheUtil.put(key, listVO);
+        }
+
+        return listVO;
+    }
+
     @Transactional(rollbackFor = Exception.class)
     @Override
     public ApiResponse<String> deleteRule(Long id) {
+        RulePO rulePO = ruleMapper.selectByPrimaryKey(id);
+        if (rulePO == null) {
+            return new ApiResponse<>(ApiResponse.FAIL.getCode(), "The rules do not exist");
+        }
+
+        List<TenantRulePO> list = tenantRuleMapper.findByCode(rulePO.getRuleCode());
+        if (CollectionUtils.isNotEmpty(list)) {
+            List<String> tenantNameList = list.stream().map(po -> tenantRepository.find(po.getTenantId()).getTenantName()).toList();
+            return new ApiResponse<>(ApiResponse.FAIL.getCode(), "Rules are selected with tenants: " + String.join(",", tenantNameList));
+        }
+
         ruleMapper.deleteByPrimaryKey(id);
         ruleScanResultMapper.deleteByRuleId(id);
         dbCacheUtil.clear(cacheKey);
@@ -291,7 +351,7 @@ public class RuleServiceImpl implements RuleService {
 
         Locale locale = LocaleContextHolder.getLocale();
         if (!locale.getLanguage().equals(Locale.CHINA.getLanguage())) {
-            for (RuleTypeVO ruleTypeVO : list){
+            for (RuleTypeVO ruleTypeVO : list) {
                 ruleTypeVO.setTypeName(RuleType.getByRuleTypeEn(ruleTypeVO.getTypeName()));
             }
         }
@@ -368,4 +428,86 @@ public class RuleServiceImpl implements RuleService {
             return ruleVO;
         }).toList();
     }
+
+    @Override
+    public synchronized ApiResponse<String> addTenantSelectRule(AddTenantSelectRuleRequest req) {
+        RulePO rulePO = ruleMapper.findOne(req.getRuleCode());
+        if (rulePO == null) {
+            return new ApiResponse<>(ApiResponse.FAIL_CODE, "The rules do not exist");
+        }
+
+        if (Status.invalid.name().equals(rulePO.getStatus())) {
+            return new ApiResponse<>(ApiResponse.FAIL_CODE, "The rules are not enabled");
+        }
+
+        Long tenantId = UserInfoContext.getCurrentUser().getUserTenantId();
+        TenantRulePO tenantRulePO = tenantRuleMapper.findOne(tenantId, req.getRuleCode());
+        if (tenantRulePO != null) {
+            return new ApiResponse<>(ApiResponse.FAIL_CODE, "The rules have been added to the optional list");
+        }
+
+        tenantRulePO = new TenantRulePO();
+        tenantRulePO.setTenantId(tenantId);
+        tenantRulePO.setRuleCode(req.getRuleCode());
+        tenantRuleMapper.insertSelective(tenantRulePO);
+
+        log.info("user:{}, addTenantSelectRule, req:{}", UserInfoContext.getCurrentUser(), req);
+
+        dbCacheUtil.clear(tenantSelectRuleCacheKey);
+
+        return ApiResponse.SUCCESS;
+    }
+
+    @Override
+    public ApiResponse<String> deleteTenantSelectRule(String ruleCode) {
+        RulePO rulePO = ruleMapper.findOne(ruleCode);
+        if (rulePO == null) {
+            return new ApiResponse<>(ApiResponse.FAIL_CODE, "The rules do not exist");
+        }
+
+        Long tenantId = UserInfoContext.getCurrentUser().getUserTenantId();
+        TenantRulePO tenantRulePO = tenantRuleMapper.findOne(tenantId, ruleCode);
+        if (tenantRulePO == null) {
+            return new ApiResponse<>(ApiResponse.FAIL_CODE, "The rules do not exist");
+        }
+
+        if (!tenantId.equals(tenantRulePO.getTenantId())) {
+            return new ApiResponse<>(ApiResponse.FAIL_CODE, "The rules do not belong to the current tenant");
+        }
+
+        tenantRuleMapper.deleteByPrimaryKey(tenantRulePO.getId());
+
+        // Delete the corresponding risk data
+        ruleScanResultMapper.deleteByRuleIdAndTenantId(rulePO.getId(), tenantId);
+
+        log.info("user:{}, deleteTenantSelectRule, ruleCode:{}", UserInfoContext.getCurrentUser(), ruleCode);
+
+        dbCacheUtil.clear(tenantSelectRuleCacheKey);
+
+        return ApiResponse.SUCCESS;
+    }
+
+    @Override
+    public ApiResponse<String> batchDeleteTenantSelectRule(List<String> ruleCodeList) {
+        for (String ruleCode : ruleCodeList) {
+            deleteTenantSelectRule(ruleCode);
+        }
+
+        return ApiResponse.SUCCESS;
+    }
+
+    @Override
+    public List<RuleVO> queryAllTenantSelectRuleList() {
+        List<RulePO> all = tenantRuleMapper.findAllList(UserInfoContext.getCurrentUser().getUserTenantId());
+        return all.stream().map(po -> {
+            RuleVO ruleVO = new RuleVO();
+            ruleVO.setRuleCode(po.getRuleCode());
+            ruleVO.setRuleName(po.getRuleName());
+            ruleVO.setPlatform(po.getPlatform());
+            ruleVO.setRuleDesc(po.getRuleDesc());
+            ruleVO.setId(po.getId());
+            return ruleVO;
+        }).toList();
+    }
+
 }
